Inline count using the old query
This is a problem if you use any of the odata operators that depend on the new aggregation stuff. 

I must admit changing the aggregation stuff like this might be considered bad form, maybe clone it instead? It probably should never be a problem however since it's the last thing that uses the aggregation object.